[PROG-9] Reorganize code
