+++ src/main.cpp
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <glog/logging.h>
 #include <SDL.h>
+#include <exception>
 #include "src/game.h"
 
 #undef main //SDL defines main and it needs to be undefined
@@ -17,7 +18,14 @@ int main(int argc, const char* argv[]) {
         LOG(ERROR) << "Could not initialize SDL! Error: " << SDL_GetError();
     }
 
-    int exit_code = Game().run();
+    int exit_code;
+
+    try{
+        exit_code = Game().run();
+    } catch(std::exception& e){
+        LOG(ERROR) << e.what();
+        exit_code = -1;
+    }
 
     SDL_Quit();
     google::ShutdownGoogleLogging();
