+++ src/pcg/agent_generators/agents/look_ahead_agent.cpp
@@ -1,125 +1,20 @@
-#include <vector>
-#include <cstdlib>
 #include <random>
 #include <glog/logging.h>
 #include <algorithm>
 #include "src/pcg/agent_generators/agents/look_ahead_agent.h"
 #include "src/pcg/utils/grid_to_entity_parser.h"
-#include "src/pcg/utils/direction.h"
-#include "src/pcg/utils/entity_creator.h"
-
-std::vector<int> fill_vector_with_range(int from, int to);
-
-template<typename T>
-class RandomIterator {
-    RandomNumberGenerator& rng;
-    std::vector<T> elements;
-    int index = 0;
-
-public:
-    explicit RandomIterator(std::vector<T>&& elements, RandomNumberGenerator& rng)
-            : rng(rng),
-              elements(std::move(elements)) {
-        shuffle();
-    }
-
-    T next(bool should_increment = true) {
-        assert((has_next()));
-
-        return elements[should_increment ? index++ : index];
-    }
-
-    bool has_next() {
-        return index < elements.size();
-    }
-
-    void reset() {
-        index = 0;
-        shuffle();
-    }
-
-
-private:
-    void shuffle() {
-        auto random_number_engine = std::default_random_engine(rng.random());
-        std::shuffle(elements.begin(), elements.end(), random_number_engine);
-    }
-
-};
-
-template<typename T, typename K>
-class RandomIteratorPair {
-    RandomIterator<T> first;
-    RandomIterator<K> second;
-
-
-public:
-    RandomIteratorPair(RandomIterator<T>&& first, RandomIterator<K>&& second)
-            : first(std::move(first)), second(std::move(second)) {}
-
-    std::pair<T, K> next() {
-        const T first_result = first.next();
-        const bool should_reset_first = !first.has_next();
-        const K second_result = second.next(should_reset_first);
-
-        if (should_reset_first) {
-            first.reset();
-        }
-
-        return std::make_pair(first_result, second_result);
-    }
-
-    bool has_next() {
-        return second.has_next();
-    }
-};
-
-class RoomIterator {
-    RandomIteratorPair<int, int> room_size_iterator;
-
-public:
-    RoomIterator(int min_size, int max_size, RandomNumberGenerator& rng)
-            : room_size_iterator(RandomIterator<int>(fill_vector_with_range(min_size, max_size), rng),
-                                 RandomIterator<int>(fill_vector_with_range(min_size, max_size), rng)) {}
-
-    Dimensions2i next() {
-        std::pair<int, int> room_size = room_size_iterator.next();
-        return {room_size.first, room_size.second};
-    }
-
-    bool has_next() {
-        return room_size_iterator.has_next();
-    }
-};
-
-class CorridorIterator {
-    RandomIteratorPair<int, Direction> corridor_iterator;
-
-public:
-    CorridorIterator(int min_size, int max_size, RandomNumberGenerator& rng)
-            : corridor_iterator(RandomIterator<int>(fill_vector_with_range(min_size, max_size), rng),
-                                RandomIterator<Direction>(std::vector(DIRECTIONS.begin(), DIRECTIONS.end()), rng)) {}
-
-    std::pair<int, Direction> next() {
-        return corridor_iterator.next();
-    }
-
-    bool has_next() {
-        return corridor_iterator.has_next();
-    }
-};
-
+#include "src/pcg/agent_generators/agents/utils/spatial_iterators.h"
 
 void LookAheadAgent::run(Scene& scene, Grid& grid, RandomNumberGenerator& rng) {
     position = get_starting_position(grid, rng);
-    //BBoxGridWrapper bbox_grid(grid);
     DLOG(INFO) << "Starting with position " << position;
 
     while (true) {
         if (try_to_place_room(scene, grid, rng)) {
         }
 
-        if (corridors_buffer.size() > MAX_CORRIDOR_SEEK_TIME || !try_to_place_corridor(scene, grid, rng)) {
+        if (corridors_buffer.size() >= MAX_CORRIDOR_SEEK_TIME
+            || !try_to_place_corridor(scene, grid, rng)) {
             break;
         }
     }
@@ -143,22 +38,6 @@ bool LookAheadAgent::try_to_place_room(Scene& scene, Grid& grid, RandomNumberGen
     return false;
 }
 
-bool LookAheadAgent::try_to_place_room(Scene& scene, BBoxGridWrapper& bbox_grid, RandomNumberGenerator& rng) {
-    RoomIterator room_iterator(MIN_ROOM_SIZE, MAX_ROOM_SIZE, rng);
-
-    while (room_iterator.has_next()) {
-        BoundingBox2i room_box = BoundingBox2i::from_dimensions_centered(position, room_iterator.next());
-
-        if (can_place_room(bbox_grid, room_box)) {
-            place_room(scene, bbox_grid, room_box);
-            return true;
-        }
-    }
-    DLOG(INFO) << "Could not place any room";
-
-    return false;
-}
-
 bool LookAheadAgent::can_place_room(Grid& grid, BoundingBox2i room_box) {
     if (!grid.contains(room_box)) {
         return false;
@@ -172,27 +51,12 @@ bool LookAheadAgent::can_place_room(Grid& grid, BoundingBox2i room_box) {
                                });
 }
 
-bool LookAheadAgent::can_place_room(BBoxGridWrapper& bbox_grid, BoundingBox2i room_box) {
-    return std::ranges::none_of(bbox_grid.begin(), bbox_grid.end(), [room_box](const GridBox& grid_box) {
-        return grid_box.box.collides_with(room_box);
-    });
-}
-
 void LookAheadAgent::place_room(Scene& scene, Grid& grid, BoundingBox2i room_box) {
     position = room_box.get_middle();
     flush_corridors_buffer(scene, grid);
 
     DLOG(INFO) << "Placed room " << room_box << " and updated position to " << position;
     DLOG(INFO) << grid << '\n';
-
-    rooms_buffer.emplace_back(room_box);
-}
-
-void LookAheadAgent::place_room(Scene& scene, BBoxGridWrapper& bbox_grid, BoundingBox2i room_box) {
-    position = room_box.get_middle();
-    flush_corridors_buffer(scene, bbox_grid);
-
-    DLOG(INFO) << "Placed room " << room_box << " and updated position to " << position;
 
     rooms_buffer.emplace_back(room_box);
 }
@@ -280,10 +144,10 @@ LookAheadAgent::get_corridor_box_with_updated_position(int length, Direction dir
 }
 
 Point2i LookAheadAgent::get_starting_position(const Grid& grid, RandomNumberGenerator& rng) const {
-    //FIXME: it ignores starting coordinates - assumes its (0,0) - cause grid model does not keep them.
+    //it ignores starting coordinates - assumes its (0,0) - cause grid model does not keep them.
     return Point2i(
-            rng.random(MIN_ROOM_SIZE, grid.get_width() - MIN_ROOM_SIZE),
-            rng.random(MIN_ROOM_SIZE, grid.get_height() - MIN_ROOM_SIZE)
+            static_cast<int>(rng.random(MIN_ROOM_SIZE, grid.get_width() - MIN_ROOM_SIZE)),
+            static_cast<int>(rng.random(MIN_ROOM_SIZE, grid.get_height() - MIN_ROOM_SIZE))
     );
 }
 
@@ -293,7 +157,6 @@ void LookAheadAgent::flush_rooms_buffer(Scene& scene, Grid& grid) {
     for (const BoundingBox2i& room_box: rooms_buffer) {
         grid.fill(room_box, GridElement::ROOM);
         DLOG(INFO) << "Added room\n" << grid << '\n';
-        //EntityCreatorImpl::create_room_floor(scene, room_box);
     }
 
     rooms_buffer.clear();
@@ -305,31 +168,7 @@ void LookAheadAgent::flush_corridors_buffer(Scene& scene, Grid& grid) {
     for (const BoundingBox2i& corridor_box: corridors_buffer) {
         grid.fill_no_override(corridor_box, GridElement::CORRIDOR);
         DLOG(INFO) << "Added room\n" << grid << '\n';
-        //EntityCreatorImpl::create_corridor_floor(scene, corridor_box);
     }
 
     corridors_buffer.clear();
-}
-
-void LookAheadAgent::flush_corridors_buffer(Scene& scene, BBoxGridWrapper& bbox_grid) {
-    DLOG(INFO) << "Flushing corridors buffer";
-
-    for (const BoundingBox2i corridor_box: corridors_buffer) {
-        bbox_grid.add(corridor_box, GridElement::CORRIDOR);
-    }
-
-    corridors_buffer.clear();
-}
-
-
-std::vector<int> fill_vector_with_range(int from, int to) {
-    int size = to - from + 1;
-    std::vector<int> result;
-    result.reserve(size);
-
-    for (int i = from; i <= to; ++i) {
-        result.push_back(i);
-    }
-
-    return result;
 }
